enable canvas-panel and image-service to rotate
This opens up the rotation parameter in Atlas to CanvasPanel and ImageService.

From what I can tell, the rotation-parameter does not operate properly on a sequence (and likely choices) because it would rotate each canvas separately… but not sure how often we have that as a use-case / need.

I _think_ i’ve wired this through properly so that any change to the `rotation` attribute should rotate the canvas.